Some cleanup

+++ src/game/mario.c
@@ -159,7 +159,7 @@ Bool32 is_anim_past_frame(struct MarioState *m, AnimFrame16 animFrame) {
  * Rotates the animation's translation into the global coordinate system
  * and returns the animation's flags.
  */
-s16 find_mario_anim_flags_and_translation(struct Object *obj, s32 yaw, Vec3s translation) {
+s32 find_mario_anim_flags_and_translation(struct Object *obj, s32 yaw, Vec3s translation) {
     struct Animation *curAnim = (void *) obj->header.gfx.animInfo.curAnim;
     AnimFrame32 animFrame  = geo_update_animation_frame(&obj->header.gfx.animInfo, NULL);
     AnimIndex  *animIndex  = segmented_to_virtual((void *) curAnim->index );
@@ -190,7 +190,7 @@ void update_mario_pos_for_anim(struct MarioState *m) {
 /**
  * Finds the vertical translation from Mario's animation.
  */
-s16 return_mario_anim_y_translation(struct MarioState *m) {
+s32 return_mario_anim_y_translation(struct MarioState *m) {
     Vec3s translation;
     find_mario_anim_flags_and_translation(m->marioObj, 0, translation);
     return translation[1];
